test: move test files into src now that they no longer reference bin
