+++ src/__tests__/cli.ts
