+++ tsconfig.json
@@ -14,5 +14,5 @@
     "declaration": true
   },
   "include": ["src", "typings"],
-  "exclude": ["node_modules", "test/**"]
+  "exclude": ["node_modules", "**/__tests__/**", "**/__mocks__/**"]
 }
